kamp-core/src/main/kotlin/ch/leadrian/samp/kamp/core/runtime/entity/property/MapObjectCoordinatesProperty.kt
package ch.leadrian.samp.kamp.core.runtime.entity.property

import ch.leadrian.samp.kamp.core.api.data.Vector3D
import ch.leadrian.samp.kamp.core.api.data.vector3DOf
import ch.leadrian.samp.kamp.core.api.entity.MapObject
import ch.leadrian.samp.kamp.core.runtime.SAMPNativeFunctionExecutor
import ch.leadrian.samp.kamp.core.runtime.types.ReferenceFloat
import kotlin.properties.ReadWriteProperty
import kotlin.reflect.KProperty

internal class MapObjectCoordinatesProperty(
        private val nativeFunctionExecutor: SAMPNativeFunctionExecutor
) : ReadWriteProperty<MapObject, Vector3D> {

    private val x = ReferenceFloat()
    private val y = ReferenceFloat()
    private val z = ReferenceFloat()

    override fun getValue(thisRef: MapObject, property: KProperty<*>): Vector3D {
        nativeFunctionExecutor.getObjectPos(objectid = thisRef.id.value, x = x, y = y, z = z)
        return vector3DOf(x = x.value, y = y.value, z = z.value)
    }

    override fun setValue(thisRef: MapObject, property: KProperty<*>, value: Vector3D) {
        nativeFunctionExecutor.setObjectPos(objectid = thisRef.id.value, x = value.x, y = value.y, z = value.z)
    }

}